Flatten file structure w/ fat modules
